+++ ui/packages/shared/client/src/parca/debuginfo/v1alpha1/debuginfo.ts
@@ -42,6 +42,12 @@ export interface ShouldInitiateUploadRequest {
      * @generated from protobuf field: parca.debuginfo.v1alpha1.DebuginfoType type = 4;
      */
     type: DebuginfoType;
+    /**
+     * Type of build ID.
+     *
+     * @generated from protobuf field: parca.debuginfo.v1alpha1.BuildIDType build_id_type = 5;
+     */
+    buildIdType: BuildIDType;
 }
 /**
  * ShouldInitiateUploadResponse is the response for ShouldInitiateUpload.
@@ -98,6 +104,12 @@ export interface InitiateUploadRequest {
      * @generated from protobuf field: parca.debuginfo.v1alpha1.DebuginfoType type = 5;
      */
     type: DebuginfoType;
+    /**
+     * Type of build ID.
+     *
+     * @generated from protobuf field: parca.debuginfo.v1alpha1.BuildIDType build_id_type = 6;
+     */
+    buildIdType: BuildIDType;
 }
 /**
  * InitiateUploadResponse is the response to an InitiateUploadRequest.
@@ -477,6 +489,37 @@ export enum DebuginfoType {
      */
     SOURCES = 2
 }
+/**
+ * BuildIDType is the type of build ID.
+ *
+ * @generated from protobuf enum parca.debuginfo.v1alpha1.BuildIDType
+ */
+export enum BuildIDType {
+    /**
+     * The build ID is unknown.
+     *
+     * @generated from protobuf enum value: BUILD_ID_TYPE_UNKNOWN_UNSPECIFIED = 0;
+     */
+    BUILD_ID_TYPE_UNKNOWN_UNSPECIFIED = 0,
+    /**
+     * The build ID is a GNU build ID.
+     *
+     * @generated from protobuf enum value: BUILD_ID_TYPE_GNU = 1;
+     */
+    BUILD_ID_TYPE_GNU = 1,
+    /**
+     * The build ID is an opaque hash.
+     *
+     * @generated from protobuf enum value: BUILD_ID_TYPE_HASH = 2;
+     */
+    BUILD_ID_TYPE_HASH = 2,
+    /**
+     * The build ID is a Go build ID.
+     *
+     * @generated from protobuf enum value: BUILD_ID_TYPE_GO = 3;
+     */
+    BUILD_ID_TYPE_GO = 3
+}
 // @generated message type with reflection information, may provide speed optimized methods
 class ShouldInitiateUploadRequest$Type extends MessageType<ShouldInitiateUploadRequest> {
     constructor() {
@@ -484,7 +527,8 @@ class ShouldInitiateUploadRequest$Type extends MessageType<ShouldInitiateUploadR
             { no: 1, name: "build_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
             { no: 2, name: "hash", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
             { no: 3, name: "force", kind: "scalar", T: 8 /*ScalarType.BOOL*/ },
-            { no: 4, name: "type", kind: "enum", T: () => ["parca.debuginfo.v1alpha1.DebuginfoType", DebuginfoType, "DEBUGINFO_TYPE_"] }
+            { no: 4, name: "type", kind: "enum", T: () => ["parca.debuginfo.v1alpha1.DebuginfoType", DebuginfoType, "DEBUGINFO_TYPE_"] },
+            { no: 5, name: "build_id_type", kind: "enum", T: () => ["parca.debuginfo.v1alpha1.BuildIDType", BuildIDType] }
         ]);
     }
     create(value?: PartialMessage<ShouldInitiateUploadRequest>): ShouldInitiateUploadRequest {
@@ -493,6 +537,7 @@ class ShouldInitiateUploadRequest$Type extends MessageType<ShouldInitiateUploadR
         message.hash = "";
         message.force = false;
         message.type = 0;
+        message.buildIdType = 0;
         if (value !== undefined)
             reflectionMergePartial<ShouldInitiateUploadRequest>(this, message, value);
         return message;
@@ -513,6 +558,9 @@ class ShouldInitiateUploadRequest$Type extends MessageType<ShouldInitiateUploadR
                     break;
                 case /* parca.debuginfo.v1alpha1.DebuginfoType type */ 4:
                     message.type = reader.int32();
+                    break;
+                case /* parca.debuginfo.v1alpha1.BuildIDType build_id_type */ 5:
+                    message.buildIdType = reader.int32();
                     break;
                 default:
                     let u = options.readUnknownField;
@@ -538,6 +586,9 @@ class ShouldInitiateUploadRequest$Type extends MessageType<ShouldInitiateUploadR
         /* parca.debuginfo.v1alpha1.DebuginfoType type = 4; */
         if (message.type !== 0)
             writer.tag(4, WireType.Varint).int32(message.type);
+        /* parca.debuginfo.v1alpha1.BuildIDType build_id_type = 5; */
+        if (message.buildIdType !== 0)
+            writer.tag(5, WireType.Varint).int32(message.buildIdType);
         let u = options.writeUnknownFields;
         if (u !== false)
             (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
@@ -611,7 +662,8 @@ class InitiateUploadRequest$Type extends MessageType<InitiateUploadRequest> {
             { no: 2, name: "size", kind: "scalar", T: 3 /*ScalarType.INT64*/, L: 0 /*LongType.BIGINT*/ },
             { no: 3, name: "hash", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
             { no: 4, name: "force", kind: "scalar", T: 8 /*ScalarType.BOOL*/ },
-            { no: 5, name: "type", kind: "enum", T: () => ["parca.debuginfo.v1alpha1.DebuginfoType", DebuginfoType, "DEBUGINFO_TYPE_"] }
+            { no: 5, name: "type", kind: "enum", T: () => ["parca.debuginfo.v1alpha1.DebuginfoType", DebuginfoType, "DEBUGINFO_TYPE_"] },
+            { no: 6, name: "build_id_type", kind: "enum", T: () => ["parca.debuginfo.v1alpha1.BuildIDType", BuildIDType] }
         ]);
     }
     create(value?: PartialMessage<InitiateUploadRequest>): InitiateUploadRequest {
@@ -621,6 +673,7 @@ class InitiateUploadRequest$Type extends MessageType<InitiateUploadRequest> {
         message.hash = "";
         message.force = false;
         message.type = 0;
+        message.buildIdType = 0;
         if (value !== undefined)
             reflectionMergePartial<InitiateUploadRequest>(this, message, value);
         return message;
@@ -644,6 +697,9 @@ class InitiateUploadRequest$Type extends MessageType<InitiateUploadRequest> {
                     break;
                 case /* parca.debuginfo.v1alpha1.DebuginfoType type */ 5:
                     message.type = reader.int32();
+                    break;
+                case /* parca.debuginfo.v1alpha1.BuildIDType build_id_type */ 6:
+                    message.buildIdType = reader.int32();
                     break;
                 default:
                     let u = options.readUnknownField;
@@ -672,6 +728,9 @@ class InitiateUploadRequest$Type extends MessageType<InitiateUploadRequest> {
         /* parca.debuginfo.v1alpha1.DebuginfoType type = 5; */
         if (message.type !== 0)
             writer.tag(5, WireType.Varint).int32(message.type);
+        /* parca.debuginfo.v1alpha1.BuildIDType build_id_type = 6; */
+        if (message.buildIdType !== 0)
+            writer.tag(6, WireType.Varint).int32(message.buildIdType);
         let u = options.writeUnknownFields;
         if (u !== false)
             (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);