+++ ui/packages/shared/profile/src/ProfileMetricsGraph/index.tsx
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import {useEffect, useMemo, useRef, useState} from 'react';
+import {useEffect, useMemo} from 'react';
 
 import {RpcError} from '@protobuf-ts/runtime-rpc';
 import {AnimatePresence, motion} from 'framer-motion';
@@ -30,6 +30,7 @@ import {capitalizeOnlyFirstLetter, getStepDuration} from '@parca/utilities';
 import {MergedProfileSelection, ProfileSelection} from '..';
 import MetricsGraph from '../MetricsGraph';
 import {useMetricsGraphDimensions} from '../MetricsGraph/useMetricsGraphDimensions';
+import useGrpcQuery from '../useGrpcQuery';
 
 interface ProfileMetricsEmptyStateProps {
   message: string;
@@ -60,7 +61,8 @@ interface ProfileMetricsGraphProps {
   profile: ProfileSelection | null;
   from: number;
   to: number;
-  timeRange: DateTimeRange;
+  sumByLoading: boolean;
+  sumBy: string[];
   setTimeRange: (range: DateTimeRange) => void;
   addLabelMatcher: (
     labels: {key: string; value: string} | Array<{key: string; value: string}>
@@ -77,7 +79,6 @@ interface ProfileMetricsGraphProps {
 export interface IQueryRangeState {
   response: QueryRangeResponse | null;
   isLoading: boolean;
-  isRefreshing?: boolean;
   error: RpcError | null;
 }
 
@@ -91,28 +92,14 @@ const getStepCountFromScreenWidth = (pixelsPerPoint: number): number => {
   return Math.round(width / pixelsPerPoint);
 };
 
-const EMPTY_SUM_BY: string[] = [];
-
 export const useQueryRange = (
   client: QueryServiceClient,
   queryExpression: string,
   start: number,
   end: number,
-  timeRange: DateTimeRange,
-  sumBy: string[] = EMPTY_SUM_BY,
+  sumBy: string[],
   skip = false
 ): IQueryRangeState => {
-  const previousQueryParams = useRef<{
-    queryExpression?: string;
-    timeRange?: DateTimeRange;
-    isRefresh?: boolean;
-  }>({});
-  const [isLoading, setLoading] = useState<boolean>(!skip);
-  const [state, setState] = useState<IQueryRangeState>({
-    response: null,
-    isLoading,
-    error: null,
-  });
   const metadata = useGrpcMetadata();
   const [stepCountStr, setStepCount] = useURLStateNew('step_count');
 
@@ -134,27 +121,11 @@ export const useQueryRange = (
     }
   }, [stepCountStr, defaultStepCount, setStepCount]);
 
-  useEffect(() => {
-    void (async () => {
-      if (skip) {
-        return;
-      }
-
-      if (
-        previousQueryParams.current.queryExpression !== queryExpression ||
-        previousQueryParams.current.timeRange !== timeRange
-      ) {
-        previousQueryParams.current.queryExpression = queryExpression;
-        previousQueryParams.current.timeRange = timeRange;
-        previousQueryParams.current.isRefresh = undefined;
-      } else {
-        previousQueryParams.current.isRefresh = true;
-      }
-
-      setLoading(true);
-
+  const {data, isLoading, error} = useGrpcQuery<QueryRangeResponse | undefined>({
+    key: ['query-range', queryExpression, start, end, (sumBy ?? []).join(','), stepCount, metadata],
+    queryFn: async () => {
       const stepDuration = getStepDuration(start, end, stepCount);
-      const call = client.queryRange(
+      const {response} = await client.queryRange(
         {
           query: queryExpression,
           start: Timestamp.fromDate(new Date(start)),
@@ -165,26 +136,16 @@ export const useQueryRange = (
         },
         {meta: metadata}
       );
+      return response;
+    },
+    options: {
+      retry: false,
+      enabled: !skip,
+      staleTime: 1000 * 60 * 5, // 5 minutes
+    },
+  });
 
-      call.response
-        .then(response => {
-          setState({response, isLoading: false, error: null});
-          setLoading(false);
-          return null;
-        })
-        .catch(error => {
-          setState({response: null, isLoading: false, error});
-          setLoading(false);
-        })
-        .finally(() => {
-          if (previousQueryParams.current.isRefresh !== undefined) {
-            previousQueryParams.current.isRefresh = undefined;
-          }
-        });
-    })();
-  }, [client, queryExpression, start, end, metadata, timeRange, sumBy, skip, stepCount]);
-
-  return {...state, isLoading, isRefreshing: previousQueryParams.current.isRefresh};
+  return {isLoading, error: error as RpcError | null, response: data ?? null};
 };
 
 const ProfileMetricsGraph = ({
@@ -197,13 +158,14 @@ const ProfileMetricsGraph = ({
   addLabelMatcher,
   onPointClick,
   comparing = false,
-  timeRange,
+  sumBy,
+  sumByLoading,
 }: ProfileMetricsGraphProps): JSX.Element => {
   const {
     isLoading: metricsGraphLoading,
     response,
     error,
-  } = useQueryRange(queryClient, queryExpression, from, to, timeRange);
+  } = useQueryRange(queryClient, queryExpression, from, to, sumBy, sumByLoading);
   const {onError, perf, authenticationErrorMessage, isDarkMode} = useParcaContext();
   const {width, height, margin, heightStyle} = useMetricsGraphDimensions(comparing);
 
@@ -224,9 +186,7 @@ const ProfileMetricsGraph = ({
   const series = response?.series;
   const dataAvailable = series !== null && series !== undefined && series?.length > 0;
 
-  if (metricsGraphLoading) {
-    return <MetricsGraphSkeleton heightStyle={heightStyle} isDarkMode={isDarkMode} />;
-  }
+  const loading = metricsGraphLoading;
 
   if (!metricsGraphLoading && error !== null) {
     if (authenticationErrorMessage !== undefined && error.code === 'UNAUTHENTICATED') {
@@ -236,52 +196,55 @@ const ProfileMetricsGraph = ({
     return <ErrorContent errorMessage={capitalizeOnlyFirstLetter(error.message)} />;
   }
 
-  if (dataAvailable) {
-    const handleSampleClick = (
-      timestamp: number,
-      _value: number,
-      labels: Label[],
-      duration: number
-    ): void => {
-      onPointClick(timestamp, labels, queryExpression, duration);
-    };
+  let sampleUnit = '';
 
-    let sampleUnit = '';
+  if (dataAvailable) {
     if (series.every((val, i, arr) => val?.sampleType?.unit === arr[0]?.sampleType?.unit)) {
       sampleUnit = series[0]?.sampleType?.unit ?? '';
     }
     if (sampleUnit === '') {
       sampleUnit = Query.parse(queryExpression).profileType().sampleUnit;
     }
+  }
 
-    return (
-      <AnimatePresence>
-        <motion.div
-          className="h-full w-full relative"
-          key="metrics-graph-loaded"
-          initial={{display: 'none', opacity: 0}}
-          animate={{display: 'block', opacity: 1}}
-          transition={{duration: 0.5}}
-        >
+  return (
+    <AnimatePresence>
+      <motion.div
+        className="h-full w-full relative"
+        key="metrics-graph-loaded"
+        initial={{display: 'none', opacity: 0}}
+        animate={{display: 'block', opacity: 1}}
+        transition={{duration: 0.5}}
+      >
+        {loading ? (
+          <MetricsGraphSkeleton heightStyle={heightStyle} isDarkMode={isDarkMode} />
+        ) : dataAvailable ? (
           <MetricsGraph
             data={series}
             from={from}
             to={to}
             profile={profile as MergedProfileSelection}
             setTimeRange={setTimeRange}
-            onSampleClick={handleSampleClick}
+            onSampleClick={(
+              timestamp: number,
+              _value: number,
+              labels: Label[],
+              duration: number
+            ): void => {
+              onPointClick(timestamp, labels, queryExpression, duration);
+            }}
             addLabelMatcher={addLabelMatcher}
             sampleUnit={sampleUnit}
             height={height}
             width={width}
             margin={margin}
           />
-        </motion.div>
-      </AnimatePresence>
-    );
-  }
-
-  return <ProfileMetricsEmptyState message="No data found. Try a different query." />;
+        ) : (
+          <ProfileMetricsEmptyState message="No data found. Try a different query." />
+        )}
+      </motion.div>
+    </AnimatePresence>
+  );
 };
 
 export default ProfileMetricsGraph;