+++ ui/packages/shared/components/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.16.293 (2024-07-10)
+
+**Note:** Version bump only for package @parca/components
+
+## 0.16.292 (2024-07-09)
+
+**Note:** Version bump only for package @parca/components
+
+## 0.16.291 (2024-07-09)
+
+**Note:** Version bump only for package @parca/components
+
 ## [0.16.290](https://github.com/parca-dev/parca/compare/@parca/components@0.16.289...@parca/components@0.16.290) (2024-07-04)
 
 **Note:** Version bump only for package @parca/components